refactor: split serve orchestration
